Skip potential duplicates when adding to filter check list
When plugins are loaded for programs that don't use event types, you
might end up with duplicate sinsp_filter_check_gen_event objects being
added to the g_filterlist filter_check_list.

This isn't harmful, but it results in duplicate entries when listing
the full set of filter fields.

Handle this by checking to see if there's already a filter_check
object with the same name (e.g. "evt") and shortdesc (e.g. "All event
types"). If there's a duplicate, simply delete the to-be-added object.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>