new(libsinsp): introduce context-free grammar for filters and top-down parser
So far, the syntax for filters didn't have a clear definition nor grammar. This constrained libsinsp
consumers, such as Falco, to define external grammars and parsers due to the impossibility of re-using
the one bundled in libsinsp. This commit introduces a context-free grammar that formally defines the
filter language. The grammar has been crafter after carefully inspecting the filters used by libs consumers
such as Falco. A top-down recursive-descent parser implementation has been added to parse filter strings with
this grammar. The parser uses the AST definition introduced in the previous commits, and is responsible for
textual-only parsing. Elaborating the AST, and compiling it in other forms (such as filterchecks), is a
responsibility of other classes.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>