refactor(libsinsp/filter): use pointers for filter AST visitor instead of references
The usage of references was not necessary for the purposes of the visitor pattern over filter ASTs, and brought
a degree of coding ambiguity due to the need of unnecessarily de-reference visitor/ast instances. This commit
changes this to just use raw pointers, which easier to read/write and more elastic.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>