fix(driver): init event_filler_arguments to fix is_socketcall
struct event_filler_arguments has a .is_socketcall member for
compatibility with those architectures having a single syscall
for all socket functions (bind, connect...).
When the architecture does NOT _HAS_SOCKETCALL though (e.g. arm64),
this field takes the value from whatever happens to be in the stack
(so likely non-zero), leading to fillers assuming it IS a socketcall
-- where in reality it is not -- and therefore exposing bogus
argument values.

The issue can be triggered by running the following on ARM:

```
telnet 127.0.0.1 8080
```

which results in sysdig reporting something like:

```
1187851 20:45:11.730173066 3 bash (1999317) > connect fd=1368509440 addr=NULL
```

Initialize the whole structure to all-zeroes so to have deterministic
behavior.

Signed-off-by: Gerlando Falauto <gerlando.falauto@sysdig.com>