Add a ast::as_string() method, using a string visitor
Add an ast::as_string() method that returns a string representation of
an ast.

It uses a new string_visitor class that builds up a string
representation of the ast while visiting. Each visitor method adds to
the string, based on the type of ast node.

as_string should also try to escape strings to reverse the string
unescaping that was done when generating the ast node. So add a new
standalone function in the libsinsp::filter namespace, in
filter/escaping.cpp, escape_str(), that does that.

The parser private method escape_str() method was misnamed--it
actually *unescapes* strings by removing quotes. So rename it to
unescape_str() and move it to a standalone libsinsp::filter function
in filter/escaping.cpp to live alongside escape_str().

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>