Allow calling stop() directly
Helgrind was reporting potential data races when calling stop() from
inside the destructor of a class that derives from
async_key_value_source. I added a lot of tracing and debugging and it
appears to be a helgrind FP. A workaround seems to be call stop()
explicitly on the derived class before calling the destructor.

This change makes the protected method public, which allows for
calling stop().

stop() is *not* required, though--stop() is still called in the
destructor as before. It will be a no-op if the background thread is
already stopped.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>