feature(sinsp-example): print info about all threads
A flag has been added to allow sinsp-example to
print information about all threads in process
and not just the main one. This may cause too
much data to be print for a regular user, but
it is useful when creating automated tests for
it.

sinsp-example now also handles SIGTERM, this is
done to simplify stopping it when running it in
a container.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>