test(parser) Add tests for saving positions in asts
Add several tests for saving positions in asts. A pos_visitor
traverses the ast and builds a string consisting only of the node
type (and, or, binary check, etc) and the current parser position.

The test cases involve parsing various expressions and ensuring that
the parse positions match the expected values. I verified all the
positions by hand.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>