feature(CI): Run e2e tests on CI
The tests have been added to an existing workflow that already builds
the userspace library and the drivers in an effort to make them a bit
faster than rebuilding everything from scratch.

Additionally, a number of mounts have been added to the
container running the tests, those are needed for everything to build
using the headers from the host and allowing the tests to insert and
remove the drivers.

The auto-generated report for the tests is automatically uploaded to GH,
for ease of debugging in case of an error.

Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>