fix(container_engine): Only return on success or all retries failed
Instead of always returning a result on the first attempt, only return
results on success or when all retries have failed.

This prevents spurious "container" events for incomplete results.

This is especially important when both docker and cri are enabled,
when both must be tried due to the cgroup pattern overlapping, but
only one actually holds the container.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>