Fix vtif/vpid fields in CLONE_20_X event built by sched_proc_fork
New sched_proc_fork tracepoint handler was created, to compensate for lack of
sys_exit tracepoint generated when clone/fork system calls return to the child
process.  The new sched_proc_fork handler builds the CLONE_20_X event which would
have been built by the missing sys_exit tracepoint.

Two of the fields in the CLONE_20_X are vtid and vpid, the thread ID and process
ID of the running thread, IN THE LOCAL PID NAMESPACE of that thread.  The new
sched_proc_fork handler has a bug -- it uses the convenience functions
task_pid_vnr() and task_tgid_vnr() to retrieve the vtid and vpid; but those
functions assume the PID namespace of the executing thread, which in the case
of the sched_proc_fork handler, is the parent thread, not the child thread.

Solution is to use task_pid_nr_ns() and task_tgid_nr_ns(), explicitly specifying
the PID namespace associated with the child thread.

Signed-off-by: Joseph Pittman <joseph.pittman@sysdig.com>