Fix extract__loginuid for COS
COS kernels have task_struct that differs from the vanilla one in
regards how audit information is stored, loginuid is present in an
intermediate structure audit_task_info [1]. This mean regular CO-RE read
will not pass verifier on COS kernels. To address it, check the field
presence first, before extracting the value. Note, it's important that
the condition is formulated this way -- the first branch should be an
existing one, otherwise it will be removed as a dead code and no
comparison will be made [2].

[1]: https://chromium.googlesource.com/chromiumos/third_party/kernel/+/096925a44076ba5c52faa84d255a847130ff341e%5E%21/#F2
[2]: https://nakryiko.com/posts/bpf-core-reference-guide/#handling-incompatible-field-and-type-changes