Allow to exclude cgroups from bpf_append_cgroup
In certain cases cgroups could be quite large, and it would become more
expensive to pass them from the kernel to user space. Allow to exclude
some cgroups if instructed to do so.

Note, that the original implementation was having dependency between cpu
& cpuacct cgroups:

    #if !IS_ENABLED(CONFIG_CGROUP_SCHED)
    #if IS_ENABLED(CONFIG_CGROUP_CPUACCT)
    	APPEND_CGROUP(cpuacct);
    #endif
    #endif

This logic would kick in only if the kernel we build on is configured to
exclude certain cgroups. Such cases don't happen in CI build pipelines,
so we got rid of them.