Delete new threadinfo if the cache is full
sinsp::remove_thread tries to resolve thread dependencies, and to be
able to do so it fetches the thread parent. Fetching the parent
considered to be an important operation, so it has query_os flag set for
get_thread_ref. Now it looks like get_thread_ref, when allocating a new
thread info, doesn't actually verify the return result of add_thread
operation. The latter might fail to add anything if the thread cache is
full, thus the following sequence can happen under load:

* We want to remove an expired thread
* It's parent got evicted from cache for whatever reason
* Falco tries to fetch the parent again via querying os
* A new threadinfo object for the parent is allocated
* Parent thread info is not added because the thread cache is full
* Parent thread info is not cleaned up afterwards, since the resulting
object is fetched again via find_thread from the cache.

As a result we've got spurious memory allocations in remove_thread.